Fix bug

+++ lampController.py
@@ -22,9 +22,9 @@ def parseTimeFromStringToInt(time):
 
 def turnLightsOn():
     for i in range(100):
-        publish.single(MQTT_PATH_RED, (i/100)*red, hostname=MQTT_SERVER)
-        publish.single(MQTT_PATH_BLUE, (i/100)*blue, hostname=MQTT_SERVER)
-        publish.single(MQTT_PATH_WHITE, (i/100)*white, hostname=MQTT_SERVER)
+        publish.single(MQTT_PATH_RED, int((i/100.0)*red), hostname=MQTT_SERVER)
+        publish.single(MQTT_PATH_BLUE, int((i/100.0)*blue), hostname=MQTT_SERVER)
+        publish.single(MQTT_PATH_WHITE, int((i/100.0)*white), hostname=MQTT_SERVER)
         print('.')
         time.sleep(0.02)
